still not have url to reset password

+++ src/main/java/com/hairsalonbookingapp/hairsalon/model/request/RequestTransaction.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class RequestTransaction {
     private String transactionType;
 
-    private long money;
+    private double money;
 
     private String description;
 }
